MAINT: refactored spectral orm submodule
